Articles/WebArticles/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "prismq-idea-sources-content-articles-webarticlesource"
version = "1.0.0"
description = "PrismQ module for gathering idea inspirations from web articles as part of the AI-powered content generation ecosystem"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "PrismQ", email = "dev@prismq.com"}
]
keywords = ["prismq", "ai", "content-generation", "web", "articles", "idea-sources"]

dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "trafilatura>=1.6.0",
    "newspaper3k>=0.2.8",
    "feedparser>=6.0.10",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "click>=8.1.7",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
web-article-source = "src.cli:main"

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["_meta/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
